[feat] 팔로워 화면 ui layer 구현
